requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
braceexpand==0.1.7
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
coremltools==7.0
cycler==0.11.0
Cython==3.0.0
datasets==2.14.5
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
hydra-core==1.3.2
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflect==7.0.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jiwer==3.0.2
joblib==1.3.2
kiwisolver==1.4.4
lazy_loader==0.3
libcst==1.0.1
librosa==0.10.0.post2
lightning-utilities==0.9.0
llvmlite==0.40.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MonkeyType==23.3.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nemo-toolkit==1.20.0
networkx==3.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pooch==1.6.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.9.6
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.4
pytz==2023.3
PyYAML==6.0.1
rapidfuzz==2.13.7
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
speechbrain==0.5.15
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.0.3
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.8
wcwidth==0.2.6
webdataset==0.2.48
Werkzeug==2.3.7
wget==3.2
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
youtokentome==1.0.6
zipp==3.16.2
